fix: running tests in paralell works now
